Build functional user sign-in/sign-out interface
- Still needs to display errors on bad requests

+++ assets/scripts/auth/api.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const app = require('../app-data').app;
+const app = require('../app-data');
 
 const signUp = (success, failure, data) => {
   $.ajax({
@@ -21,15 +21,17 @@ const signIn = (success, failure, data) => {
 };
 
 const signOut = (success, failure) => {
-  // if(!app.user) bad;
-  $.ajax({
-    method: 'DELETE',
-    url: app.api + '/sign-out/' + app.user.id,
-    headers: {
-      Authorization: 'Token token=' + app.user.token,
-    }
-  }).done(success)
-  .fail(failure);
+  console.log(app);
+  if(app.user) {
+    $.ajax({
+      method: 'DELETE',
+      url: app.api + '/sign-out/' + app.user.id,
+      headers: {
+        Authorization: 'Token token=' + app.user.token,
+      }
+    }).done(success)
+    .fail(failure);
+  }
 };
 
 module.exports = {
